+++ app/src/main/java/ru/skillbranch/skillarticles/data/repositories/ArticleRepository.kt
@@ -4,6 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.map
 import ru.skillbranch.skillarticles.data.*
+import ru.skillbranch.skillarticles.ui.custom.markdown.MarkdownElement
+import ru.skillbranch.skillarticles.ui.custom.markdown.MarkdownParser
 
 interface  IArticleRepository{
     fun loadArticleContent(articleId: String): LiveData<List<MarkdownElement>?>
