+++ app/src/main/java/ru/skillbranch/skillarticles/extensions/Mappers.kt
@@ -45,3 +45,13 @@ fun User.asMap(): Map<String, Any?> = mapOf(
     "respect"  to respect,
     "about"  to about
 )
+
+fun List<Pair<Int, Int>>.groupByBounds(bounds: List<Pair<Int, Int>>): List<MutableList<Pair<Int, Int>>> {
+    return bounds.fold(mutableListOf<MutableList<Pair<Int, Int>>>()){acc, pair ->
+        val res = this.filter {
+            it.first >= pair.first && it.second <= pair.second
+        }.toMutableList()
+        acc.add(res)
+        acc
+    }
+}
