+++ app/src/main/java/ru/skillbranch/skillarticles/ui/custom/markdown/IMarkdownView.kt
@@ -3,8 +3,8 @@ package ru.skillbranch.skillarticles.ui.custom.markdown
 import android.text.Spannable
 import android.text.SpannableString
 import androidx.core.text.getSpans
-import ru.skillbranch.skillarticles.ui.custom.spans.SearchFocusSpan
-import ru.skillbranch.skillarticles.ui.custom.spans.SearchSpan
+import ru.skillbranch.skillarticles.ui.custom.markdown.spans.SearchFocusSpan
+import ru.skillbranch.skillarticles.ui.custom.markdown.spans.SearchSpan
 
 interface IMarkdownView {
     var fontSize: Float
