+++ app/src/main/java/ru/skillbranch/skillarticles/ui/custom/markdown/MarkdownBuilder.kt
@@ -11,11 +11,9 @@ import androidx.core.content.ContextCompat
 import androidx.core.text.buildSpannedString
 import androidx.core.text.inSpans
 import ru.skillbranch.skillarticles.R
-import ru.skillbranch.skillarticles.data.repositories.Element
-import ru.skillbranch.skillarticles.data.repositories.MarkdownElement
 import ru.skillbranch.skillarticles.extensions.attrValue
 import ru.skillbranch.skillarticles.extensions.dpToPx
-import ru.skillbranch.skillarticles.ui.custom.spans.*
+import ru.skillbranch.skillarticles.ui.custom.markdown.spans.*
 
 class MarkdownBuilder(context: Context) {
     private val strikeWidth: Float = context.dpToPx(4)
