+++ app/src/main/java/ru/skillbranch/skillarticles/ui/custom/markdown/MarkdownTextView.kt
@@ -1,5 +1,6 @@
 package ru.skillbranch.skillarticles.ui.custom.markdown
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.graphics.Canvas
 import android.graphics.Rect
@@ -13,29 +14,31 @@ import ru.skillbranch.skillarticles.R
 import ru.skillbranch.skillarticles.extensions.attrValue
 import ru.skillbranch.skillarticles.extensions.dpToIntPx
 import ru.skillbranch.skillarticles.extensions.dpToPx
-
+@SuppressLint("ViewConstructor")
 class MarkdownTextView constructor(
     context: Context,
     fontSize: Float,
     private val isSizeDepend: Boolean = true
 ) : AppCompatTextView(context, null, 0), IMarkdownView {
-    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
-    val searchPadding = context.dpToIntPx(56)
-    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
-    var searchBgHelper = SearchBgHelper(context) { top, bottom ->
-        focusRect.set(0, top - searchPadding, width, bottom + searchPadding)
-        requestRectangleOnScreen(focusRect, false)
-    }
+
     override var fontSize: Float = fontSize
         set(value) {
             textSize = value
             field = value
         }
+
     override val spannableContent: Spannable
         get() = text as Spannable
-    private val color = context.attrValue(R.attr.colorOnBackground, true)
 
+    private val color = context.attrValue(R.attr.colorOnBackground)
     private val focusRect = Rect()
+    private val searchPadding = context.dpToIntPx(56)
+
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    var searchBgHelper = SearchBgHelper(context) { top, bottom ->
+        focusRect.set(0, top - searchPadding, width, bottom + searchPadding)
+        requestRectangleOnScreen(focusRect, false)
+    }
 
     init {
         setTextColor(color)
@@ -45,7 +48,7 @@ class MarkdownTextView constructor(
 
     override fun onDraw(canvas: Canvas) {
         if (text is Spanned && layout != null) {
-            canvas.withTranslation(totalPaddingLeft.toFloat(), totalPaddingRight.toFloat()) {
+            canvas.withTranslation(totalPaddingLeft.toFloat(), totalPaddingTop.toFloat()) {
                 searchBgHelper.draw(canvas, text as Spanned, layout)
             }
         }
@@ -56,6 +59,4 @@ class MarkdownTextView constructor(
         if (isSizeDepend) setLineSpacing(context.dpToPx(if (size == 14f) 8 else 10), 1f)
         super.setTextSize(size)
     }
-
-
 }
