+++ app/src/main/java/ru/skillbranch/skillarticles/ui/custom/markdown/SearchBgHelper.kt
@@ -12,9 +12,9 @@ import androidx.core.graphics.ColorUtils
 import androidx.core.text.getSpans
 import ru.skillbranch.skillarticles.R
 import ru.skillbranch.skillarticles.extensions.*
-import ru.skillbranch.skillarticles.ui.custom.spans.HeaderSpan
-import ru.skillbranch.skillarticles.ui.custom.spans.SearchFocusSpan
-import ru.skillbranch.skillarticles.ui.custom.spans.SearchSpan
+import ru.skillbranch.skillarticles.ui.custom.markdown.spans.HeaderSpan
+import ru.skillbranch.skillarticles.ui.custom.markdown.spans.SearchFocusSpan
+import ru.skillbranch.skillarticles.ui.custom.markdown.spans.SearchSpan
 
 @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
 class SearchBgHelper(
