+++ app/src/main/java/ru/skillbranch/skillarticles/ui/custom/markdown/spans/SearchFocusSpan.kt
@@ -1,0 +1,5 @@
+package ru.skillbranch.skillarticles.ui.custom.markdown.spans
+
+class SearchFocusSpan() : SearchSpan() {
+
+}
