+++ app/src/main/java/ru/skillbranch/skillarticles/ui/custom/markdown/spans/SearchSpan.kt
@@ -1,0 +1,8 @@
+package ru.skillbranch.skillarticles.ui.custom.markdown.spans
+
+import android.graphics.Color
+import android.text.style.ForegroundColorSpan
+
+open class SearchSpan() : ForegroundColorSpan(Color.WHITE){
+
+}
