--- app/src/main/java/ru/skillbranch/skillarticles/ui/custom/spans/SearchFocusSpan.kt
@@ -1,7 +1,0 @@
-package ru.skillbranch.skillarticles.ui.custom.spans
-
-import android.text.TextPaint
-
-class SearchFocusSpan() : SearchSpan() {
-
-}
