--- app/src/main/java/ru/skillbranch/skillarticles/ui/custom/spans/SearchSpan.kt
@@ -1,11 +1,0 @@
-package ru.skillbranch.skillarticles.ui.custom.spans
-
-import android.graphics.Color
-import android.text.TextPaint
-import android.text.style.BackgroundColorSpan
-import android.text.style.ForegroundColorSpan
-import androidx.core.graphics.ColorUtils
-
-open class SearchSpan() : ForegroundColorSpan(Color.WHITE){
-
-}
